Made new component which renders the donation progress bar for members
