+++ src/Shared/components/DonationProgressBar/DonationProgressBar.scss
@@ -1,0 +1,22 @@
+.member-progress-bar {
+	$bar-border-radius: 5px;
+	width: 100%;
+	height: 10px;
+	border-radius: $bar-border-radius;
+	background-color: #ddd;
+	overflow: hidden;
+	margin-bottom: 15px;
+
+	.progress-bar-container{
+		display: flex;
+		position: relative;
+		z-index: 1;
+
+		.bar-fill {
+			position: absolute;
+			height: 10px;
+			border-top-right-radius: $bar-border-radius;
+			border-bottom-right-radius: $bar-border-radius;
+		}
+	}
+}
